requirements/base.txt
Django==1.6.5
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
South==1.0.0
psycopg2==2.5.3
celery==3.1.17
django-tastypie==0.12.1